Move McpTransport into src/Contracts/Mcp
Every other interface in the project lives under src/Contracts (with
subnamespaces for Gateway, Providers, Files). McpTransport was the only
interface declared outside of it.
